requirements.txt
numpy
matplotlib
tensorflow-gpu
tensorflow_datasets
Pillow
scipy
tensorflow_hub==0.4.0
